DeviceSettings: add an option to enable Samsung Keyboard dock
Our sec_keyboard driver does not properly unregister the input device when no
dock is detected, which makes the system believe that a physical keyboard is
present at all times, which prevents the software keyboard from being shown.

Some user own a Keyboard-Dock and meight miss the ability to use the dock if
it gets disabled completely.
It's build as module to give the user the ability to load the driver manually.

Todo: UI needs definately an update here, SwitchPreference ?

Change-Id: Ie351f18b996972595e9198481bd568bdd8f8adc6